+++ common/src/main/java/io/github/flemmli97/simplequests/JsonCodecs.java
@@ -40,7 +40,7 @@ public class JsonCodecs {
                     Codec.INT.fieldOf("z").forGetter(Vec3i::getZ)
             ).apply(instance, BlockPos::new));
 
-    public static Codec<NumberProvider> NUMER_PROVIDER_CODEC = JsonCodecs.jsonCodecBuilder(GSON::toJsonTree, e -> GSON.fromJson(e, NumberProvider.class), "NumberProvider");
+    public static Codec<NumberProvider> NUMBER_PROVIDER_CODEC = JsonCodecs.jsonCodecBuilder(GSON::toJsonTree, e -> GSON.fromJson(e, NumberProvider.class), "NumberProvider");
 
     public static <E> Codec<List<Either<E, Pair<E, String>>>> optionalDescriptiveList(Codec<E> codec, String error) {
         return nonEmptyList(Codec.either(codec, Codec.STRING.dispatch("description", Pair::getSecond, e -> Codec.pair(codec, Codec.unit(e)))), error);